app/requirements.txt
asgiref==3.7.2
astroid==2.13.5
attrs==23.1.0
autopep8==2.0.4
beautifulsoup4==4.12.2
bleach==6.0.0
soupsieve==2.5
click==8.1.7
Django==4.2.5
django-cors-headers==4.2.0
django-environ==0.11.2
django-filter==23.2
djangorestframework==3.14.0
django-parler==2.3
django-parler-rest==2.2
django-nested-admin==4.0.2
drf-jwt==1.19.2
drf-nested-routers==0.93.4
drf-yasg==1.21.7
drf-writable-nested==0.7.0
gunicorn==21.2.0
iniconfig==2.0.0
isort==5.12.0
lazy-object-proxy==1.9.0
Markdown==3.4.4
mccabe==0.7.0
packaging==23.1
pluggy==1.3.0
psycopg2==2.9.7
py==1.11.0
pycodestyle==2.11.0
pylint==2.15.10
pyparsing==3.1.1
pytest==7.4.1
pytest-drf==1.1.3
pytest-django==4.5.2
pytest-mock==3.11.1
pytest-xdist==3.3.1 
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
toml==0.10.2
wrapt==1.15.0
whitenoise==6.5.0
dadata==21.10.1
pexpect==4.8.0
pur==7.3.1
pandas==1.5.3
google-api-python-client==2.98.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
pygsheets==2.0.6
django-spaghetti-and-meatballs==0.4.2
Unidecode==1.3.6
sentry-sdk==1.30.0
django-simple-history==3.3.0
openpyxl==3.1.2
xlrd==2.0.1
html2text==2020.1.16
six==1.16.0
webencodings==0.5.1
redbox==0.2.1
coreapi==2.3.3
